Port tests/unittest to Google Test; fix all compile errors
Conversion (UnitTest++ → GTest):
  - CHECK() → EXPECT_TRUE()
  - SUITE(S)/TEST(T) → TEST(S, T)
  - TEST_FIXTURE(F, T) → TEST_F(F, T)
  - SessionFixture/ClientFixture inherit ::testing::Test
  - main.cc: gtest InitGoogleTest/RUN_ALL_TESTS
  - CMakeLists.txt: aerie_unittests target with GTest::GTest

Compile fixes:
  - file.test.cc: stray } between suites removed; File* fp = nullptr
  - namespace.test.cc: PseudoDirInode missing return_dentry() stub added
  - client.fixture.h: AllocateContainer returns E_SUCCESS; unnamed params
  - session.fixture.h: AllocateContainer returns E_SUCCESS; unnamed params
  - bitset.test.cc: int→size_t comparison
  - CMakeLists: -Wno-unused-parameter -Wno-unused-variable for test mocks
  - libfs/CMakeLists: BUILD_TESTS option (OFF by default)
  - docker/Dockerfile.ci: libgtest-dev added

+++ tests/unittest/common/bitset.test.cc
@@ -6,7 +6,7 @@ bool CompareBitSet(DynamicBitSet* bset, std::set<int> bitpos)
 {
     std::vector<int>::iterator iter;
 
-    for (int i = 0; i < bset->Size(); i++)
+    for (size_t i = 0; i < (size_t)bset->Size(); i++)
     {
         if (bset->IsSet(i) && bitpos.find(i) == bitpos.end())
         {
